inference/pyproject.toml
[project]
name = "inference"
version = "0.1.0"
description = "推理服务"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi >=0.115.12,<0.116.0",
    "uvicorn >=0.34.2,<0.35.0",
    "flagembedding >=1.3.4,<2.0.0",
    "aiohttp >=3.12.0,<4.0.0",
    "pydantic >=2.11.5,<3.0.0",
    "python-multipart >=0.0.20,<0.0.21",
    "nltk >=3.9.1,<4.0.0",
    "marker-pdf[full] >=1.7.3,<2.0.0",
    "transformers ==4.51.2"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"